requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.45.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.6.3
idna==3.2
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
